backend/account/src/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
grpcio==1.50.0
grpcio-tools==1.50.0
Pillow==9.3.0
protobuf==4.21.9
psycopg2-binary==2.9.5
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
